Make tabs scrollable
This commit makes the tabs in the main screen scrollable by replacing `TabRow` with `ScrollableTabRow`.
