--- app/src/main/java/com/example/hashcatgui/HashtopolisApiModels.kt
@@ -1,11 +1,0 @@
-package com.example.hashcatgui
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class Agent(
-    val id: Int,
-    val name: String,
-    val status: String,
-    val lastActivity: String
-)
