--- app/src/main/java/com/example/hashcatgui/MainActivity.kt
@@ -1,33 +1,0 @@
-package com.example.hashcatgui
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.viewModels
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.ui.Modifier
-import com.example.hashcatgui.ui.theme.HashcatGUITheme
-
-class MainActivity : ComponentActivity() {
-    private val mainViewModel: MainViewModel by viewModels {
-        MainViewModel.MainViewModelFactory(application, HashcatApiClient())
-    }
-    private val hashtopolisViewModel: HashtopolisViewModel by viewModels()
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            HashcatGUITheme {
-                // A surface container using the 'background' color from the theme
-                Surface(
-                    modifier = Modifier.fillMaxSize(),
-                    color = MaterialTheme.colorScheme.background
-                ) {
-                    MainScreen(mainViewModel, hashtopolisViewModel)
-                }
-            }
-        }
-    }
-}
