--- app/src/main/java/com/example/hashcatgui/ui/tabs/WordlistTab.kt
@@ -1,32 +1,0 @@
-package com.example.hashcatgui.ui.tabs
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import com.example.hashcatgui.MainViewModel
-
-@Composable
-fun WordlistTab(viewModel: MainViewModel) {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(16.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center
-    ) {
-        OutlinedTextField(
-            value = viewModel.wordlistPath.value,
-            onValueChange = { viewModel.wordlistPath.value = it },
-            label = { Text("Remote Wordlist Path") },
-            modifier = Modifier.fillMaxWidth()
-        )
-    }
-}
