--- app/src/main/java/com/example/hashcatgui/ui/theme/Color.kt
@@ -1,24 +1,0 @@
-package com.example.hashcatgui.ui.theme
-
-import androidx.compose.ui.graphics.Color
-
-val Purple80 = Color(0xFFD0BCFF)
-val PurpleGrey80 = Color(0xFFCCC2DC)
-val Pink80 = Color(0xFFEFB8C8)
-
-val Purple40 = Color(0xFF6650a4)
-val PurpleGrey40 = Color(0xFF625b71)
-val Pink40 = Color(0xFF7D5260)
-
-// Expressive Color Palette
-val Primary = Color(0xFF6200EE)
-val PrimaryVariant = Color(0xFF3700B3)
-val Secondary = Color(0xFF03DAC6)
-val SecondaryVariant = Color(0xFF018786)
-val Background = Color(0xFF121212)
-val Surface = Color(0xFF121212)
-val OnPrimary = Color.White
-val OnSecondary = Color.Black
-val OnBackground = Color.White
-val OnSurface = Color.White
-val Error = Color(0xFFCF6679)
