+++ app/src/main/java/com/hereliesaz/halfhashedkitty/HashtopolisViewModel.kt
@@ -1,0 +1,28 @@
+package com.hereliesaz.halfhashedkitty
+
+import androidx.compose.runtime.mutableStateListOf
+import androidx.compose.runtime.mutableStateOf
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import kotlinx.coroutines.launch
+
+class HashtopolisViewModel(private val apiClient: HashtopolisApiClient = HashtopolisApiClient()) : ViewModel() {
+
+    val serverUrl = mutableStateOf("")
+    val apiKey = mutableStateOf("")
+    val agents = mutableStateListOf<Agent>()
+    val errorMessage = mutableStateOf<String?>(null)
+
+    fun getAgents() {
+        viewModelScope.launch {
+            try {
+                errorMessage.value = null
+                val agentList = apiClient.getAgents(serverUrl.value, apiKey.value)
+                agents.clear()
+                agents.addAll(agentList)
+            } catch (e: Exception) {
+                errorMessage.value = "Error fetching agents: ${e.message}"
+            }
+        }
+    }
+}
