+++ app/src/main/res/xml/backup_rules.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<full-backup-content>
+    <!-- For more information on backup rules, see https://developer.android.com/guide/topics/data/autobackup -->
+    <include domain="file" path="."/>
+    <include domain="database" path="."/>
+    <include domain="sharedpref" path="."/>
+    <include domain="external" path="."/>
+    <include domain="root" path="."/>
+</full-backup-content>
